Update ch01.asciidoc
Hi, new to both bitcoin and github so apols if comments have not been recorded in a useful format, or are themselves not useful comments. I have reviewed mainly to increase my own understanding and both enjoyed reading the chapter and thought it was very informative. I have prefaced my comments in the text with 'JB' so I hope that makes them easy to find. General observation would be that I am not sure I could have picked it up cold, but perhaps your intention is not for complete beginners anyway. Will continue now with the subsequent chapters... Kind regards, James