Change spelling
I think it is better to say 'a RETURN output' than 'an RETURN output' because RETURN is a word, not an acronym, and begins with a consonant sound.

If it was an acronym then then it would make sense to say 'an R.E.T.U.R.N. output' because it would start with a vowel sound. If it was OP_RETURN then it would also make sense to say 'an OP_RETURN output' because it would start with a vowel sound.