Made sentence clearer
Changed `but` to `more` in the sentence `Cryptography means secret writing but than just encryption.`
Not sure if this was the author's intent.
How about using the following sentence instead and removing the reference to encryption since it's referred to a bit later?
`...extensively in computer security. We usually talk about cryptography in the context of secret communications. Cryptography can also be used...`