Discerning Bitcoin from bitcoin
After working on the chainspl.it project (with Eric Lombrozo, Adam Back, JW Weatherman et al.) It became clear to me that the holistic product/service/project “Bitcoin” is represented differently, in written communication, in contrast to (actual rights to) a single “bitcoin” (or part there of). The difference is the capitalisation of b.