Transaction fees are not competitive (ch. 2)
There is nothing in the Bitcoin protocol itself that makes the fees competitive. [A while ago I created a question on BSE](https://bitcoin.stackexchange.com/questions/100750/are-bitcoin-transaction-fees-competitive) to clarify this issue. This PR is my attempt to let the reader know that although miners on average prefer high-fee transactions, it is with respect to hashrate that they compete.