Clarify validation as exclusive to node operators
It is a common misconception that miners validate transactions. While technically correct as most miners run full nodes, the mining process does not include the process of validating transactions. For example, miners who operate via pools are reliant on the full nodes operated by pools and therefore reliant on the validation processes of pool operators. In a recent discussion on this topic parts of this chapter have been quoted in reference to miners validating blocks, specifically "Miners validate new transactions and record them on the global ledger". This PR aims to clarify that validating transactions is not part of the mining process but part of full node operations.