Update ch05.asciidoc
In the section on extended keys, the line "Because the extended key is 512 or 513 bits, it is also much longer than other Base58Check-encoded strings we have seen previously" mentions that the extended key is 513 bits. But how is it 513 bits? Shouldn't it instead be as follows:
Because the extended key is 512 bits [256 bits(private key) + 256 bits(chain code)] or 520 bits [264 bits(public key) + 256 bits(chain code)], it is also much longer than other Base58Check-encoded strings we have seen previously.