CH01::Receving: remove misleading claim about address identity association
Original text claimed that there is no association between a user's
identity and their addresses prior to using that address in most
wallets, but many lightweight wallets send even their not-yet-used
addresses to remote servers to scan for incoming payments.  This is
often done over IPv4 or IPv6, which creates a link between the user's
connection and their addresses.  Update text by simply dropping this
claim.