CH02::invoice: replace "payment request" with "invoice"
The payment request language is confusing because BIP70 payment requests
look a lot like the URI used here but this is not a BIP70 payment
request (and BIP70 is dead, so we need not describe it).

Instead, update the text to use the more generic word "invoice", which
has the advantage of also being simpler English.