CH12: s/Smores signature/Foocoin/
"Smore" sounds like "Schnorr", which could be confusing.  Also, we
have upgraded signatures now without a soft fork, so the example is
confusing.  Although we could add another coin with a soft fork (sort
of), it seems a lot less likely to me, so I think it's a better cadidate
for this section.